Revert "Bug fixing+.env creation"
