+++ ConcordiaMaps/.gitignore
@@ -31,7 +31,6 @@ yarn-error.*
 
 # local env files
 .env*.local
-.env
 
 # typescript
 *.tsbuildinfo
