+++ ConcordiaMaps/app.json
@@ -1,7 +1,6 @@
 {
   "expo": {
     "name": "ConcordiaMaps",
-    "permissions": ["LOCATION"],
     "slug": "ConcordiaMaps",
     "version": "1.0.0",
     "orientation": "portrait",
@@ -20,8 +19,7 @@
       "adaptiveIcon": {
         "foregroundImage": "./assets/adaptive-icon.png",
         "backgroundColor": "#ffffff"
-      },
-      "package": "com.anonymous.ConcordiaMaps"
+      }
     },
     "web": {
       "favicon": "./assets/favicon.png"
