+++ ConcordiaMaps/components/Header.js
@@ -1,6 +1,5 @@
 import React from "react";
-import { SafeAreaView, View, Text, Image } from "react-native";
-import styles from "../styles";
+import { SafeAreaView, StyleSheet, View, Text, Image } from "react-native";
 
 function Header() {
   return (
@@ -15,5 +14,30 @@ function Header() {
     </SafeAreaView>
   );
 }
+
+const styles = StyleSheet.create({
+  header: {
+    width: "100%",
+    flexDirection: "row",
+    alignItems: "center",
+    justifyContent: "center",
+    backgroundColor: "#912338",
+    padding: 10,
+    position: "relative",
+    zIndex: 1, // Ensures the logo appears on top
+  },
+  logo: {
+    width: 50,
+    height: 50,
+    marginLeft: 20,
+  },
+  headerText: {
+    fontWeight: "bold",
+    fontFamily: "Times New Roman",
+    fontSize: 25,
+    color: "white",
+    marginLeft: 10,
+  },
+});
 
 export default Header;
