--- ConcordiaMaps/data/coloringData.js
@@ -1,1737 +1,0 @@
-export const coloringData = {
-  // SGW Campus buildings latitude and longitude
-
-  type: "FeatureCollection",
-  features: [
-    {
-      type: "Feature",
-      properties: {
-        name: "Webster Library",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57963046920737, 45.49775169163101],
-            [-73.57955946057372, 45.49771311467592],
-            [-73.57939004616142, 45.49788606249953],
-            [-73.57946464721739, 45.4979242039621],
-            [-73.57963046920737, 45.49775169163101],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "B Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57804418393232, 45.497262054608626],
-            [-73.57857121520847, 45.49671648133665],
-            [-73.57852644807298, 45.496692233512874],
-            [-73.57855662819655, 45.49666013973581],
-            [-73.5776893698917, 45.49624439390405],
-            [-73.57745131973333, 45.4964875991073],
-            [-73.57768570376251, 45.49660222699427],
-            [-73.57765113021173, 45.496649125155216],
-            [-73.57755715858981, 45.4966102278421],
-            [-73.57728270069995, 45.49689030577156],
-            [-73.57804418393232, 45.497262054608626],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "CL Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57963954438469, 45.49416226808111],
-            [-73.57933212931339, 45.49396315642389],
-            [-73.57928647247152, 45.49397603369914],
-            [-73.57923563787365, 45.49400059011248],
-            [-73.57892680209324, 45.49425807479241],
-            [-73.57927682490606, 45.4944713651569],
-            [-73.57963954438469, 45.49416226808111],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "CI Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.5800136660169, 45.49740550774803],
-            [-73.57997255671032, 45.49738739328407],
-            [-73.5799720148879, 45.49737354789099],
-            [-73.57995010580561, 45.49736224181777],
-            [-73.57992032934399, 45.49736316405077],
-            [-73.57975475733517, 45.497540250480256],
-            [-73.57984196947653, 45.49758369659381],
-            [-73.5800136660169, 45.49740550774803],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "D Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57931528402645, 45.49784992112666],
-            [-73.57941175370958, 45.49775085416516],
-            [-73.57933567826632, 45.49771696348674],
-            [-73.57924146452564, 45.49781569275376],
-            [-73.57931528402645, 45.49784992112666],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "EN Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57978236441589, 45.49671469561062],
-            [-73.57973236081729, 45.49668903210875],
-            [-73.5796213214164, 45.49680759037534],
-            [-73.57958731263017, 45.496790034446434],
-            [-73.57948439255013, 45.49689729780195],
-            [-73.57956717565388, 45.49693867667153],
-            [-73.57978236441589, 45.49671469561062],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "ER Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.58036392411584, 45.496253295070375],
-            [-73.58009133299879, 45.49616836335687],
-            [-73.5800687611461, 45.496188272035056],
-            [-73.58000217773727, 45.49616149887413],
-            [-73.57969501984768, 45.49648436691098],
-            [-73.57999578025023, 45.49662763208909],
-            [-73.58036392411584, 45.496253295070375],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "ES Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.58001199448555, 45.496119712454714],
-            [-73.57994450376027, 45.49609618044062],
-            [-73.5797769874875, 45.4962623173725],
-            [-73.57984312945435, 45.49629464859922],
-            [-73.58001199448555, 45.496119712454714],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "ET Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57994589909113, 45.496095547998095],
-            [-73.57985811101334, 45.49606208298414],
-            [-73.5797026727586, 45.49621785706964],
-            [-73.57970403526383, 45.49623069019029],
-            [-73.57977437867135, 45.49625966920917],
-            [-73.57994589909113, 45.496095547998095],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "EV Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57871965203222, 45.49562737503871],
-            [-73.57865698613432, 45.49559590636082],
-            [-73.5786779575786, 45.49555374259356],
-            [-73.57804430805658, 45.49524939986793],
-            [-73.57802539177649, 45.49525885972457],
-            [-73.57786646520893, 45.49518946849801],
-            [-73.57762463247447, 45.49543420834931],
-            [-73.57774906683876, 45.4955007912626],
-            [-73.57773506320439, 45.49551475586756],
-            [-73.57775891990988, 45.49552808014505],
-            [-73.57770099304601, 45.495594596520846],
-            [-73.57754267319861, 45.49552257542959],
-            [-73.57724191544224, 45.49583904921232],
-            [-73.57771042178814, 45.496052415748494],
-            [-73.57800219068075, 45.49574979176313],
-            [-73.57830994175416, 45.495893382805804],
-            [-73.57832459677877, 45.49588100746891],
-            [-73.57843061395349, 45.49592684274228],
-            [-73.57871965203222, 45.49562737503871],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "FA Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57959162558426, 45.49677984106009],
-            [-73.57951848892013, 45.4967460088574],
-            [-73.57942173283787, 45.49684384490757],
-            [-73.57949693176229, 45.49688106273277],
-            [-73.57959162558426, 45.49677984106009],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Faubourg Tower",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57800385438442, 45.494678406278695],
-            [-73.5774914099835, 45.49438908445663],
-            [-73.57721754501493, 45.49463246492846],
-            [-73.57744661831677, 45.49474596926282],
-            [-73.57746349352365, 45.494735221679775],
-            [-73.57752617116284, 45.494773118420056],
-            [-73.57776266198056, 45.49488353142647],
-            [-73.57800385438442, 45.494678406278695],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Faubourg Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57903715164632, 45.493812243240455],
-            [-73.57870814718952, 45.49361735395871],
-            [-73.57775091839422, 45.49444163147791],
-            [-73.57769129407737, 45.49441019233939],
-            [-73.57763180458575, 45.49441650504295],
-            [-73.57759137392902, 45.494438690372334],
-            [-73.57801687942256, 45.49467892569342],
-            [-73.57903715164632, 45.493812243240455],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Grey Nuns Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57835637694289, 45.49385136002991],
-            [-73.57826330765914, 45.49379557718456],
-            [-73.57792622825401, 45.494082355076785],
-            [-73.57786652080753, 45.494056780357624],
-            [-73.5776352666439, 45.49429512234934],
-            [-73.57774372444237, 45.49434838262121],
-            [-73.57797168572806, 45.49412150810696],
-            [-73.57801713333492, 45.49414622585567],
-            [-73.57835637694289, 45.49385136002991],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Guy-De Maisonneuve Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57911565353642, 45.495781893428074],
-            [-73.5790866367708, 45.49576919926534],
-            [-73.57906488097775, 45.49578762725386],
-            [-73.57874036161634, 45.495639381409575],
-            [-73.57842981609606, 45.49595448693759],
-            [-73.57877793844457, 45.49611572301441],
-            [-73.57911565353642, 45.495781893428074],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Grey Nuns Building - Main Entrance",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.5774835902767, 45.493199703433845],
-            [-73.57756009038253, 45.49311591030448],
-            [-73.57734404270634, 45.49299959938867],
-            [-73.57740202353776, 45.4929406899206],
-            [-73.57733311878076, 45.49290374908034],
-            [-73.57737218089878, 45.49286581152211],
-            [-73.57730625083126, 45.49283677726362],
-            [-73.5772842176633, 45.492876457546316],
-            [-73.57720999179436, 45.49284253778805],
-            [-73.5771479522582, 45.492909176533686],
-            [-73.57695338712816, 45.4928123947318],
-            [-73.57694843496124, 45.4927789802185],
-            [-73.57702613773309, 45.49272404884201],
-            [-73.57694320400022, 45.49268427266125],
-            [-73.57692862594281, 45.492692880806544],
-            [-73.5768737578814, 45.49276168347939],
-            [-73.57654588504916, 45.492603352383355],
-            [-73.57640902820658, 45.49274202207076],
-            [-73.57674194518694, 45.49290521383347],
-            [-73.57674710734163, 45.49294596639825],
-            [-73.57679691145768, 45.492958494114845],
-            [-73.57682584575862, 45.492939569188025],
-            [-73.57704456297836, 45.49303987376851],
-            [-73.57673125465115, 45.493347550081324],
-            [-73.57670924951519, 45.4933356819418],
-            [-73.57655877068407, 45.49348513630579],
-            [-73.57659371528909, 45.49350720799913],
-            [-73.5762777854248, 45.49381996270523],
-            [-73.57585403388686, 45.49361641379815],
-            [-73.5758341086693, 45.49363516073467],
-            [-73.57580471953318, 45.49361669566355],
-            [-73.57570687171757, 45.49371171539485],
-            [-73.5757480506002, 45.49372253468226],
-            [-73.57574029621996, 45.49373883554631],
-            [-73.57615433718244, 45.49394429003459],
-            [-73.57605142537771, 45.49405441768505],
-            [-73.57609198466135, 45.49408141820794],
-            [-73.57604548675951, 45.494134349689574],
-            [-73.57611452638285, 45.49416779301433],
-            [-73.57612380810366, 45.49414620297057],
-            [-73.57624458978252, 45.49421659127171],
-            [-73.57627195237843, 45.49418333304453],
-            [-73.576143971196, 45.49413053814823],
-            [-73.57617110843184, 45.49410628274251],
-            [-73.57621437875383, 45.49413581600882],
-            [-73.57627036194445, 45.49407842608571],
-            [-73.57635896258755, 45.49412139914938],
-            [-73.57643127884931, 45.49407346951983],
-            [-73.57708737162528, 45.494387459847076],
-            [-73.57735441646115, 45.49413908267624],
-            [-73.57739869805705, 45.49415289250484],
-            [-73.57742641535377, 45.49412814334556],
-            [-73.57738271583824, 45.49411192050525],
-            [-73.57741086084725, 45.494084906487416],
-            [-73.57745957974153, 45.49412260409886],
-            [-73.57760078009935, 45.49397796436645],
-            [-73.57738553031781, 45.49387969666708],
-            [-73.57705160951988, 45.49419397061726],
-            [-73.57689500991965, 45.49411872573526],
-            [-73.57691497468016, 45.49410234964401],
-            [-73.57659328083152, 45.49395546463353],
-            [-73.57648469029826, 45.49391830601786],
-            [-73.57663899725085, 45.493896559556305],
-            [-73.57682320735064, 45.4939858605091],
-            [-73.5768149476697, 45.4940161372181],
-            [-73.57687716854643, 45.494047499071286],
-            [-73.57690649922955, 45.49402992350588],
-            [-73.57698692938914, 45.49405530618009],
-            [-73.57714304543424, 45.493914365995636],
-            [-73.57677759908586, 45.49372858573494],
-            [-73.57672334073706, 45.49377189642385],
-            [-73.57665205373564, 45.49373350712233],
-            [-73.57663365366443, 45.49375163445404],
-            [-73.57671509184496, 45.49378072989893],
-            [-73.5766557043936, 45.493841602965574],
-            [-73.57649065609357, 45.493881429937204],
-            [-73.57650803416476, 45.493836876044156],
-            [-73.57653155609476, 45.49384492468647],
-            [-73.57662335961197, 45.49374813516491],
-            [-73.57665607134508, 45.49372274525933],
-            [-73.57663521784951, 45.49371231768367],
-            [-73.57679321865186, 45.493563955554066],
-            [-73.5768953309694, 45.493616934346164],
-            [-73.57684019393066, 45.4936748453793],
-            [-73.57686871265778, 45.493690134713574],
-            [-73.57686663573193, 45.493736353852825],
-            [-73.57692724925212, 45.49374013059648],
-            [-73.57694930965033, 45.493720187449554],
-            [-73.57710804663922, 45.49379149237686],
-            [-73.57715483432997, 45.49373566832509],
-            [-73.57719928249425, 45.49374268751217],
-            [-73.57725916177397, 45.493676146229205],
-            [-73.57724632613409, 45.49364575711124],
-            [-73.57730580733217, 45.49360370454798],
-            [-73.57713725386463, 45.493543207704136],
-            [-73.57712931715497, 45.49347520713235],
-            [-73.57705506926592, 45.493498094878106],
-            [-73.57688697113312, 45.493432696262545],
-            [-73.5769612903004, 45.49336027052482],
-            [-73.57699754015987, 45.49336989728121],
-            [-73.57716159181379, 45.49320274100859],
-            [-73.57712137811978, 45.493181790208496],
-            [-73.5772055029299, 45.493119076891986],
-            [-73.57742469285381, 45.49322905816561],
-            [-73.57749663269026, 45.493192553023164],
-            [-73.5774835902767, 45.493199703433845],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "GS Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.58129466483888, 45.496770541808644],
-            [-73.58139760976722, 45.4966677975338],
-            [-73.5812410667171, 45.496610052046236],
-            [-73.58114574454318, 45.49671406892233],
-            [-73.58129466483888, 45.496770541808644],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Henry F. Hall",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57900728513012, 45.497709976128675],
-            [-73.57954257102988, 45.49716511478255],
-            [-73.5788327353804, 45.49681927191142],
-            [-73.57829046749058, 45.497370661897946],
-            [-73.57900728513012, 45.497709976128675],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "K Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57939273225148, 45.49788208257124],
-            [-73.57963203345119, 45.49763906829227],
-            [-73.57956180417747, 45.4976042826064],
-            [-73.5794863632123, 45.497681196939226],
-            [-73.57953013777126, 45.49770491398078],
-            [-73.5794793657305, 45.49775716703789],
-            [-73.5794617415673, 45.4977510457596],
-            [-73.57943975161962, 45.497770103907555],
-            [-73.57940890052078, 45.49776373168905],
-            [-73.57931878516523, 45.497853083538416],
-            [-73.57939273225148, 45.49788208257124],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "LC Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57742126584051, 45.49671045279831],
-            [-73.57738606846488, 45.49669444769944],
-            [-73.57742067275117, 45.496648651022554],
-            [-73.57739333453988, 45.496631562766794],
-            [-73.57719081889063, 45.496840917238416],
-            [-73.57725775030482, 45.49687562170158],
-            [-73.57742126584051, 45.49671045279831],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "LD Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57744562036864, 45.49657291147142],
-            [-73.57737024089097, 45.49654288257929],
-            [-73.57723444757293, 45.49668260856731],
-            [-73.57722532346848, 45.49668180358219],
-            [-73.57710572143947, 45.49680108458915],
-            [-73.57711722600371, 45.49680540229036],
-            [-73.57709606137104, 45.49682822302637],
-            [-73.57713006098541, 45.49684064487644],
-            [-73.5771438803163, 45.49684318647351],
-            [-73.57717981092581, 45.4968175771761],
-            [-73.57719785413902, 45.49682630827965],
-            [-73.57744562036864, 45.49657291147142],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Learning Square",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57955713111238, 45.49653030013195],
-            [-73.5796681923123, 45.49641352707994],
-            [-73.57957431915551, 45.4963625546406],
-            [-73.57966422584043, 45.49626617026453],
-            [-73.5794725130558, 45.496174419599356],
-            [-73.57925435712785, 45.49639313811011],
-            [-73.57955713111238, 45.49653030013195],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "M Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.5797304436249, 45.49741830445336],
-            [-73.57982837568642, 45.49732291296064],
-            [-73.57980290069625, 45.49730624336195],
-            [-73.57979904278098, 45.497295933479336],
-            [-73.57978123813972, 45.49728412918904],
-            [-73.5797628286834, 45.4972880724294],
-            [-73.57966591774161, 45.49738800994453],
-            [-73.5797304436249, 45.49741830445336],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "MK Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57920678074862, 45.496465269681664],
-            [-73.57915228205708, 45.49643861397229],
-            [-73.57907457709064, 45.49651625135664],
-            [-73.57905899359793, 45.49650893854232],
-            [-73.57895728993046, 45.496615849517525],
-            [-73.57902962443804, 45.49665044578251],
-            [-73.57920678074862, 45.496465269681664],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "MM Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.576448205863, 45.4946251295863],
-            [-73.57636331436997, 45.49460367283589],
-            [-73.5763032151996, 45.49467548143386],
-            [-73.5763185094921, 45.494681437535206],
-            [-73.5762890081115, 45.49471239119171],
-            [-73.57629947994316, 45.49472782365979],
-            [-73.57622179963168, 45.49481193850417],
-            [-73.57626441863589, 45.49482932966498],
-            [-73.576448205863, 45.4946251295863],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "MN Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57636294405725, 45.49460044654367],
-            [-73.57630028550314, 45.49455478203791],
-            [-73.57614981899344, 45.49473057863395],
-            [-73.57619187035942, 45.49475234402777],
-            [-73.57627026213848, 45.49466206692139],
-            [-73.57630606510136, 45.49467407776396],
-            [-73.57636294405725, 45.49460044654367],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Montefiore Building, MT Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57621646987235, 45.49452153944455],
-            [-73.57599075734316, 45.494422094415825],
-            [-73.57573353862857, 45.49470310762965],
-            [-73.57589513920679, 45.494764985949985],
-            [-73.57601375229858, 45.49463492652609],
-            [-73.57609075921414, 45.494666458911894],
-            [-73.57621646987235, 45.49452153944455],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "John Molson School Of Business",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57955980449731, 45.4952819786065],
-            [-73.57930730462672, 45.49512504817298],
-            [-73.57923695682284, 45.495189381863895],
-            [-73.57879389257668, 45.49495285928191],
-            [-73.57847192366296, 45.495203852764746],
-            [-73.57907380119016, 45.49547883720072],
-            [-73.57908815851162, 45.49546493361481],
-            [-73.57922462832728, 45.49553597700179],
-            [-73.57955980449731, 45.4952819786065],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "MI Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.58250764624533, 45.49084431720334],
-            [-73.58240944896465, 45.49078885662763],
-            [-73.58228186602918, 45.49090234018598],
-            [-73.58237408405249, 45.490953409922156],
-            [-73.58250764624533, 45.49084431720334],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "MU Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57970094803, 45.49779373223507],
-            [-73.57963968818981, 45.49775370959796],
-            [-73.57947467303703, 45.49792517999211],
-            [-73.57953817887179, 45.49795945671269],
-            [-73.57970094803, 45.49779373223507],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "P Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57926011451251, 45.49662364146198],
-            [-73.57919278613511, 45.49659349815843],
-            [-73.57910970787542, 45.49668425320482],
-            [-73.57917456382931, 45.49671977200927],
-            [-73.57926011451251, 45.49662364146198],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "PR Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57981889711098, 45.49703285946589],
-            [-73.57996772575335, 45.49688521114848],
-            [-73.57999075054482, 45.496890192572636],
-            [-73.58004482085487, 45.49682605293992],
-            [-73.57995655886491, 45.49678620996929],
-            [-73.57989898158893, 45.49684712406622],
-            [-73.57995083703219, 45.496876297386564],
-            [-73.57990476237553, 45.49691728106683],
-            [-73.57985609802759, 45.496894931946485],
-            [-73.57975252311752, 45.497008458908596],
-            [-73.57981889711098, 45.49703285946589],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Q Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57918709372808, 45.496578646665284],
-            [-73.57913721953837, 45.49655614535979],
-            [-73.5790401772967, 45.496650456711734],
-            [-73.579093394481, 45.49668008984892],
-            [-73.57918709372808, 45.496578646665284],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "R Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57951001343068, 45.49674195048678],
-            [-73.57942438903574, 45.49670423007544],
-            [-73.57933756579936, 45.49680091465976],
-            [-73.5794156147027, 45.49684070670443],
-            [-73.57951001343068, 45.49674195048678],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "RR Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57932959560091, 45.496799648120145],
-            [-73.57946157167204, 45.49665365960715],
-            [-73.57938330757514, 45.496616216692956],
-            [-73.57925139383369, 45.496764895208514],
-            [-73.57932959560091, 45.496799648120145],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "S Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57991786128142, 45.4973613925433],
-            [-73.57988452176264, 45.497355074097754],
-            [-73.57988210111363, 45.49733571240924],
-            [-73.57985854479462, 45.49732162657952],
-            [-73.57984103012426, 45.4973220215777],
-            [-73.57974699744538, 45.49742225932283],
-            [-73.57975872982028, 45.49743284819118],
-            [-73.57973808227416, 45.49745853772018],
-            [-73.57980221943687, 45.4974894290348],
-            [-73.57991786128142, 45.4973613925433],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Samuel Bronfman Building, SB Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.58625998751384, 45.49646581748095],
-            [-73.58585028265352, 45.496498657738755],
-            [-73.58586290161153, 45.496519916814435],
-            [-73.585919386565, 45.49655144765944],
-            [-73.58589687242828, 45.496575152289836],
-            [-73.5858511373377, 45.49655573229728],
-            [-73.58583885143305, 45.49657874799635],
-            [-73.5860622862493, 45.496674595135374],
-            [-73.58608754413488, 45.49664697808373],
-            [-73.58610316110064, 45.49665903877133],
-            [-73.58610372053285, 45.49668791565898],
-            [-73.58614672234664, 45.49670284555165],
-            [-73.58622185052114, 45.496610255641514],
-            [-73.58620646793577, 45.49658741068296],
-            [-73.5862369576097, 45.4965883204537],
-            [-73.58624843550227, 45.496567875071634],
-            [-73.58618951045787, 45.49655671237983],
-            [-73.58619194682707, 45.49653655406138],
-            [-73.5862692593503, 45.49651934985283],
-            [-73.58625998751384, 45.49646581748095],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "T Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57933905241327, 45.49665684080165],
-            [-73.57926706068656, 45.496622250865215],
-            [-73.57918248486236, 45.49671882146899],
-            [-73.57924983081199, 45.49675389290428],
-            [-73.57933905241327, 45.49665684080165],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Toronto-Dominion Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.5784598933456, 45.49510789036992],
-            [-73.57833834487586, 45.49503500148748],
-            [-73.57830512454898, 45.4950533487939],
-            [-73.57826196665651, 45.49503037674512],
-            [-73.57829198065758, 45.49500883538869],
-            [-73.57815200340639, 45.49493267256298],
-            [-73.57805569340616, 45.495017229890806],
-            [-73.57839492498172, 45.495165233094156],
-            [-73.5784598933456, 45.49510789036992],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "V Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.5800068745401, 45.4969729548786],
-            [-73.57993340074958, 45.49694005504551],
-            [-73.5798281434358, 45.49704939644889],
-            [-73.57989409936914, 45.49707851765152],
-            [-73.5800068745401, 45.4969729548786],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Visual Art Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57374969408997, 45.49540303302143],
-            [-73.57350901436469, 45.495644073786394],
-            [-73.57382724840083, 45.49577848425636],
-            [-73.57355281499174, 45.496057343935036],
-            [-73.5737850873306, 45.49616399388691],
-            [-73.57428219421034, 45.4956531280576],
-            [-73.57374969408997, 45.49540303302143],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "X Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57983697716314, 45.49690466071752],
-            [-73.5796851169702, 45.49682991965673],
-            [-73.57959619954758, 45.496928355987905],
-            [-73.57975245235463, 45.49699669460031],
-            [-73.57983697716314, 45.49690466071752],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Z Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.57983377945916, 45.49690572647475],
-            [-73.57975771972154, 45.496867362146304],
-            [-73.57967374402948, 45.496958080042106],
-            [-73.57975246499761, 45.496994262759415],
-            [-73.57983377945916, 45.49690572647475],
-          ],
-        ],
-      },
-    },
-    // Loyola Campus buildings latitude and longitude
-    {
-      type: "Feature",
-      properties: {
-        name: "Administration Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.64012899431809, 45.457835010047376],
-            [-73.63983999983382, 45.45772097818761],
-            [-73.63982835994423, 45.457732066521345],
-            [-73.63980930910019, 45.457737417311535],
-            [-73.63978502055787, 45.45776582973994],
-            [-73.63979382110945, 45.457778969730896],
-            [-73.63978151301146, 45.45779344710388],
-            [-73.6398268350443, 45.45780896958519],
-            [-73.63982204705665, 45.45782087450996],
-            [-73.63985491261913, 45.457835927676655],
-            [-73.6397573670367, 45.45794162124062],
-            [-73.63973991053294, 45.457945036266835],
-            [-73.6396929497692, 45.4579299102563],
-            [-73.6396532544876, 45.457983431482035],
-            [-73.63969748999784, 45.4579971410202],
-            [-73.63970176671245, 45.458012136187165],
-            [-73.63971334738365, 45.45801666574454],
-            [-73.63961629340398, 45.458128234282015],
-            [-73.63959388969722, 45.45811227255817],
-            [-73.63957872035535, 45.458122686016964],
-            [-73.63953393774409, 45.45810870928406],
-            [-73.63950153566243, 45.4581256081313],
-            [-73.63949162440727, 45.45815585611851],
-            [-73.63948435863405, 45.45817613182675],
-            [-73.63959994631531, 45.458221418453746],
-            [-73.63959247141662, 45.4582417314021],
-            [-73.63964133906707, 45.45826247088141],
-            [-73.63965839513457, 45.458245729850205],
-            [-73.63977580822275, 45.45829152208506],
-            [-73.63983246052966, 45.45822080449622],
-            [-73.63975651771189, 45.45819311073197],
-            [-73.63985643105123, 45.458062852310235],
-            [-73.63987837052498, 45.45807252167924],
-            [-73.63992522466327, 45.458015287961246],
-            [-73.63990726887694, 45.458003374249245],
-            [-73.64001025664263, 45.45787692292683],
-            [-73.640076239391, 45.45790331406215],
-            [-73.64012899431809, 45.457835010047376],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "BB Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.63924850034303, 45.459719761772675],
-            [-73.63917538077546, 45.45969247910722],
-            [-73.63910048734955, 45.459793049414856],
-            [-73.63917192711203, 45.45982080061529],
-            [-73.63924850034303, 45.459719761772675],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "BH Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.63917385206783, 45.459691924098735],
-            [-73.63909628749916, 45.459663329098845],
-            [-73.63902047565499, 45.45976351521213],
-            [-73.63909733010927, 45.45979112251831],
-            [-73.63917385206783, 45.459691924098735],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Central Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.63992048036631, 45.45802990898943],
-            [-73.63989639132535, 45.45806093633615],
-            [-73.63996101904729, 45.458086928255625],
-            [-73.63992068030005, 45.458142246486545],
-            [-73.64015930815559, 45.45822579144138],
-            [-73.64014938823092, 45.458243241299385],
-            [-73.64044799377567, 45.458356389706154],
-            [-73.64043966410982, 45.45836581559823],
-            [-73.64067205440426, 45.45845982612417],
-            [-73.64070697047045, 45.45840892029085],
-            [-73.64077639246058, 45.458432212535286],
-            [-73.64082305259134, 45.458376597204904],
-            [-73.64075313899303, 45.45834948548483],
-            [-73.64079328457174, 45.45830789152254],
-            [-73.64056274044194, 45.458216943565674],
-            [-73.64056058322303, 45.458198630143926],
-            [-73.64027553115595, 45.45808880076763],
-            [-73.64026806017759, 45.458099395982885],
-            [-73.6400228333605, 45.45800469470029],
-            [-73.640004372169, 45.45803027088502],
-            [-73.63997090396484, 45.458016426659356],
-            [-73.63994859112654, 45.458041056244326],
-            [-73.63992048036631, 45.45802990898943],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Communication Studies and Journalism Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.64064294316807, 45.45766754878201],
-            [-73.6405459567221, 45.45763306156505],
-            [-73.64075944910564, 45.45736136805547],
-            [-73.64063230896187, 45.4573075245645],
-            [-73.6406445928628, 45.457282884512296],
-            [-73.64039736518431, 45.45717797998972],
-            [-73.64020436428909, 45.4574161928052],
-            [-73.64007849292767, 45.457365292226854],
-            [-73.64004508146434, 45.45730887275434],
-            [-73.64006884532523, 45.45730200483848],
-            [-73.64002166308596, 45.45722065988747],
-            [-73.63994374930361, 45.457219983911756],
-            [-73.63987478072906, 45.45723627234895],
-            [-73.63981694598174, 45.45727128400682],
-            [-73.63977479981322, 45.457315700621564],
-            [-73.63976863000319, 45.45740329329291],
-            [-73.63982667225929, 45.457476095677094],
-            [-73.63994459304621, 45.45746258564148],
-            [-73.6399390009149, 45.45744255389499],
-            [-73.64002933683946, 45.45743148600678],
-            [-73.64015580607033, 45.45748774616027],
-            [-73.64005158449126, 45.45762031235685],
-            [-73.64032086830642, 45.45772326124464],
-            [-73.6402952681371, 45.45776313429823],
-            [-73.64050999883878, 45.457839933399725],
-            [-73.64064294316807, 45.45766754878201],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Stinger Dome",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.63708397577038, 45.45738094341158],
-            [-73.63636967869348, 45.4569597960857],
-            [-73.63523248517185, 45.45791905102695],
-            [-73.63596137210425, 45.45834213584445],
-            [-73.63633934641047, 45.45801350822538],
-            [-73.63643927198386, 45.45807968697062],
-            [-73.63647759606003, 45.45804397231959],
-            [-73.6363858185982, 45.45797791250462],
-            [-73.63692532575001, 45.45751322287492],
-            [-73.63702283624811, 45.45756412966449],
-            [-73.63705732741845, 45.457541352981735],
-            [-73.63696845419966, 45.45748742922237],
-            [-73.63708397577038, 45.45738094341158],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "F.C. Smith Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.63960567813767, 45.45865526385933],
-            [-73.63967506895716, 45.45857601821257],
-            [-73.63965988988468, 45.45854727720641],
-            [-73.63958605991282, 45.45851707061655],
-            [-73.63959068374758, 45.45850551058291],
-            [-73.6395653506561, 45.458492508194524],
-            [-73.6395847975428, 45.458463718877454],
-            [-73.6395786223281, 45.45844910258137],
-            [-73.63955691658172, 45.45844012030969],
-            [-73.63952882804931, 45.45844640612788],
-            [-73.63949965410052, 45.45843155440653],
-            [-73.63948274354411, 45.45845025065529],
-            [-73.63941944882686, 45.458424136502714],
-            [-73.63940990694627, 45.458423154842706],
-            [-73.63940487886428, 45.45842968537917],
-            [-73.63915849072937, 45.458340319917454],
-            [-73.63905260786382, 45.45847678945799],
-            [-73.63905993269739, 45.45848057277365],
-            [-73.63905239331766, 45.45849186451329],
-            [-73.6390585685442, 45.458512653429864],
-            [-73.63909274973942, 45.45852646848016],
-            [-73.63912395110682, 45.45851795428973],
-            [-73.6391290876789, 45.458510711839864],
-            [-73.63929331898693, 45.45857064495266],
-            [-73.63928822991248, 45.458580569019205],
-            [-73.63947528310175, 45.45864629725503],
-            [-73.6394877501092, 45.45862936250151],
-            [-73.63956976230978, 45.458660209453825],
-            [-73.63960567813767, 45.45865526385933],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Centre for Structural and Functional Genomics",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.64056651434012, 45.45716534498723],
-            [-73.64073210156481, 45.45695424944759],
-            [-73.64065794751792, 45.45692071229874],
-            [-73.64067997077285, 45.456895272872],
-            [-73.64063941130908, 45.456877169180785],
-            [-73.64061465964897, 45.45690104525664],
-            [-73.64037144457878, 45.45680102479329],
-            [-73.64017646543836, 45.457039169102586],
-            [-73.64044143680644, 45.45713915031331],
-            [-73.64045364490028, 45.457119632093196],
-            [-73.64056651434012, 45.45716534498723],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Hingston Hall, wing HA",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.64152391178881, 45.4594557028056],
-            [-73.64155790566805, 45.45939045306051],
-            [-73.64152947338414, 45.45937771924119],
-            [-73.64150875256526, 45.459404993232795],
-            [-73.64114825182033, 45.45926599898013],
-            [-73.64112726525047, 45.45927638876029],
-            [-73.64108991839923, 45.45925734152101],
-            [-73.64107013086574, 45.4592766878508],
-            [-73.64104218057602, 45.45926613023494],
-            [-73.64089877311945, 45.45945949518341],
-            [-73.64091019552855, 45.45946529499764],
-            [-73.6408932686348, 45.45949092540664],
-            [-73.64094529886205, 45.4595100883611],
-            [-73.64094223496252, 45.45952840339149],
-            [-73.64127765425624, 45.459655472000236],
-            [-73.64129129482355, 45.45963971982928],
-            [-73.64135017851909, 45.459667560694584],
-            [-73.64135885966365, 45.459644302030796],
-            [-73.64137562328078, 45.45964710212175],
-            [-73.64152391178881, 45.4594557028056],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Hingston Hall, wing HB",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.64229097589806, 45.45912105903963],
-            [-73.64223636843455, 45.459101631762564],
-            [-73.64224337011989, 45.4590866527227],
-            [-73.6419101845365, 45.45895439427517],
-            [-73.64189473014164, 45.45896739481978],
-            [-73.64184690508878, 45.458947990420256],
-            [-73.64182070606262, 45.45897315345118],
-            [-73.64180300916102, 45.458964707744684],
-            [-73.64165517937222, 45.45915881875632],
-            [-73.64166362461417, 45.459168210989446],
-            [-73.64165456231298, 45.45918435851337],
-            [-73.64170198360391, 45.4592088304523],
-            [-73.64169341708882, 45.45922386451926],
-            [-73.64163323296299, 45.459205266460486],
-            [-73.6415169465114, 45.45936876491302],
-            [-73.64199178634772, 45.45955379447125],
-            [-73.64199659709394, 45.45953395591741],
-            [-73.64192839723636, 45.45951001212924],
-            [-73.64206048107312, 45.459336284047964],
-            [-73.64210377851239, 45.459357557770744],
-            [-73.64212117798981, 45.459340285497774],
-            [-73.64213186643961, 45.4593468608403],
-            [-73.6422872604576, 45.45915356131346],
-            [-73.6422680857986, 45.459143844949125],
-            [-73.64229097589806, 45.45912105903963],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Hingston Hall, wing HC",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.64225415109048, 45.459626832228054],
-            [-73.64201493727222, 45.45952697131503],
-            [-73.64188676485406, 45.45969828325602],
-            [-73.64191244688199, 45.459709336444064],
-            [-73.64183329036082, 45.459815838076764],
-            [-73.64200254276105, 45.4598816766661],
-            [-73.64208294392733, 45.45977566509444],
-            [-73.64212609493094, 45.45978532933171],
-            [-73.64225415109048, 45.459626832228054],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Applied Science Hub, HU Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.6421001505528, 45.45843152779133],
-            [-73.64168797116206, 45.45827841898736],
-            [-73.64168527937404, 45.45830877938732],
-            [-73.64148912221236, 45.45857037751486],
-            [-73.64187773739721, 45.45872031940996],
-            [-73.6421001505528, 45.45843152779133],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Jesuit Residence",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.64331548047808, 45.45852862684387],
-            [-73.64338107669136, 45.45848278173503],
-            [-73.64334706286768, 45.458465853494914],
-            [-73.64339541856884, 45.458406549541905],
-            [-73.64332899449994, 45.458371552709934],
-            [-73.64333640584337, 45.45835769870649],
-            [-73.64324520683056, 45.45832064969028],
-            [-73.64321824619121, 45.458335217852635],
-            [-73.64316305438015, 45.4583071653303],
-            [-73.64309898805357, 45.45836561321559],
-            [-73.64312717325298, 45.458387167879835],
-            [-73.64308378041665, 45.45844511414731],
-            [-73.64314437768392, 45.45847073967818],
-            [-73.64314251815001, 45.458491606297116],
-            [-73.64321862147054, 45.458523932321356],
-            [-73.64323145884954, 45.45850530960092],
-            [-73.64331548047808, 45.45852862684387],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "PB Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.63853480614185, 45.45689395939047],
-            [-73.63824344897839, 45.45677609462675],
-            [-73.63803642823315, 45.457045858296425],
-            [-73.63832959426888, 45.45715213674748],
-            [-73.63853480614185, 45.45689395939047],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Perform Center",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.637852388541, 45.456998855441014],
-            [-73.63699719753899, 45.45667180020601],
-            [-73.63678630487718, 45.45693543803068],
-            [-73.63764221434359, 45.45726758557831],
-            [-73.637852388541, 45.456998855441014],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Physical Service",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.63945117922626, 45.45928868259156],
-            [-73.63922907530134, 45.459566553107265],
-            [-73.63935460241588, 45.459626851468045],
-            [-73.6393356720333, 45.45966313905987],
-            [-73.64013587087622, 45.45994343854127],
-            [-73.6403191709627, 45.45970666840367],
-            [-73.64013291455542, 45.45964417449568],
-            [-73.64015296666703, 45.459605511626194],
-            [-73.63967062062294, 45.459414123305805],
-            [-73.63964665619253, 45.459451705367286],
-            [-73.63954223990443, 45.459409780832004],
-            [-73.6395711208874, 45.45933700021661],
-            [-73.63945117922626, 45.45928868259156],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Psychology Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.64080499972603, 45.458800428198714],
-            [-73.64044751355917, 45.45865747152172],
-            [-73.64042291931153, 45.45869144420675],
-            [-73.64040390341918, 45.4586896941513],
-            [-73.6402343663103, 45.45891812998241],
-            [-73.64024830457716, 45.45892401944286],
-            [-73.64014482827798, 45.45904528355348],
-            [-73.64055564519343, 45.459205224023066],
-            [-73.64061753605365, 45.459139893615344],
-            [-73.64055937351927, 45.459110767392076],
-            [-73.64079697396164, 45.45880480066532],
-            [-73.64080499972603, 45.458800428198714],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Oscar Peterson Concert Hall, PT Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.63921188778131, 45.45932474638235],
-            [-73.63878296459295, 45.45916881016055],
-            [-73.63866322722025, 45.459320656563676],
-            [-73.63912929778407, 45.4594968035737],
-            [-73.63924028294348, 45.45936654290904],
-            [-73.63921258615237, 45.45934968260798],
-            [-73.63921188778131, 45.45932474638235],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Recreational and Athletics",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.63833447731767, 45.45716439501069],
-            [-73.63852738730273, 45.45688475743003],
-            [-73.63824030461045, 45.45677385354915],
-            [-73.63818323309073, 45.45683422542248],
-            [-73.638060131625, 45.45678878595672],
-            [-73.638140371926, 45.45668816140599],
-            [-73.63736051424974, 45.45639031690288],
-            [-73.63735326595369, 45.45640351346543],
-            [-73.63733292577349, 45.45639215991943],
-            [-73.63699544569536, 45.45666659728039],
-            [-73.63785354228679, 45.45699662298844],
-            [-73.63782740602268, 45.45704857100162],
-            [-73.63785581671911, 45.45705660686269],
-            [-73.63795007752185, 45.456964493927615],
-            [-73.63805161163017, 45.457003394307094],
-            [-73.63802142863983, 45.457053236440174],
-            [-73.63833447731767, 45.45716439501069],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Loyola Jesuit Hall and Conference Center",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.64130685160704, 45.45849294973755],
-            [-73.6408984845733, 45.458354743662426],
-            [-73.64090297615358, 45.45834182534114],
-            [-73.64087329684291, 45.45833092596169],
-            [-73.640858885649, 45.45833396477187],
-            [-73.64076719786149, 45.458459295562136],
-            [-73.64088906263198, 45.458506759133314],
-            [-73.64069441953367, 45.45875426060053],
-            [-73.64073790749231, 45.4587737529736],
-            [-73.64076030300699, 45.45874861521827],
-            [-73.64105361171215, 45.45885952954464],
-            [-73.64130685160704, 45.45849294973755],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Solar House, SH Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.64262036433873, 45.459309764073595],
-            [-73.64261232704165, 45.45922263069764],
-            [-73.64243509713249, 45.45922644547639],
-            [-73.64243762712708, 45.45931174505452],
-            [-73.64262036433873, 45.459309764073595],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Student Center",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.63931594977598, 45.45929594000398],
-            [-73.63940637962706, 45.45917455512782],
-            [-73.63938871224275, 45.45916423084148],
-            [-73.63942522986424, 45.45911654592041],
-            [-73.63911038954912, 45.45899289095345],
-            [-73.63904390106296, 45.45908346879017],
-            [-73.63899387175562, 45.45907139971183],
-            [-73.63895678248807, 45.459132352966265],
-            [-73.6389920169611, 45.4591470982603],
-            [-73.6389379884532, 45.459224950636695],
-            [-73.63920525516396, 45.45931658032717],
-            [-73.6392194149963, 45.459289842956196],
-            [-73.63924582606751, 45.45930307831176],
-            [-73.63926391768786, 45.459278699054835],
-            [-73.63931594977598, 45.45929594000398],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Richard J. Renaud Science Complex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.64060744972808, 45.457115838109985],
-            [-73.64057127090942, 45.457165571006044],
-            [-73.64147402288205, 45.45752250718539],
-            [-73.64125404757964, 45.45779289020905],
-            [-73.64122801338183, 45.45778855879249],
-            [-73.64117889473302, 45.457859307878834],
-            [-73.64119975600728, 45.45786956691467],
-            [-73.64117142122001, 45.457902381130936],
-            [-73.64112623420205, 45.45788299357966],
-            [-73.64106092619117, 45.45797226868061],
-            [-73.64109966235054, 45.45799318658533],
-            [-73.64085400648548, 45.45831399770955],
-            [-73.64093318764986, 45.45834653960259],
-            [-73.64105437489063, 45.45819873795446],
-            [-73.64142951616398, 45.45834809009841],
-            [-73.6416900588656, 45.458010088090475],
-            [-73.64165636636018, 45.45799605984081],
-            [-73.64175638169945, 45.457882997095],
-            [-73.6418041878684, 45.457900993906094],
-            [-73.64203083597157, 45.45759172093145],
-            [-73.64191763168918, 45.45754911147469],
-            [-73.64200012222287, 45.457429676338705],
-            [-73.6414074141029, 45.45720694783944],
-            [-73.64142857688589, 45.457185126988605],
-            [-73.64139663628166, 45.45717000460081],
-            [-73.64137286865261, 45.45718584915363],
-            [-73.6413119873122, 45.45715523088461],
-            [-73.64129045157966, 45.45717342870296],
-            [-73.64126841160285, 45.45716352908687],
-            [-73.64126733177022, 45.45714578767581],
-            [-73.6409957538524, 45.457040310777586],
-            [-73.64101221929326, 45.45701759112595],
-            [-73.64095510996887, 45.45699654920833],
-            [-73.64093214847455, 45.45702437509806],
-            [-73.64082590093172, 45.456978119631884],
-            [-73.64069960114027, 45.45714112662603],
-            [-73.64060744972808, 45.457115838109985],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Terrebonne Building, TA Building",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.64096928054643, 45.459960845116],
-            [-73.64087516369197, 45.45992174505636],
-            [-73.64079872101141, 45.4600108112237],
-            [-73.64089367203884, 45.46004867276724],
-            [-73.64096928054643, 45.459960845116],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "TB Annex",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.64096928054643, 45.459960845116],
-            [-73.64087516369197, 45.45992174505636],
-            [-73.64079872101141, 45.4600108112237],
-            [-73.64089367203884, 45.46004867276724],
-            [-73.64096928054643, 45.459960845116],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Vanier Library",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.63893471101957, 45.459219544549114],
-            [-73.63898901198758, 45.459150435578294],
-            [-73.638959861861, 45.45913330145774],
-            [-73.63897385821676, 45.45909146687049],
-            [-73.6382772768585, 45.45881487698162],
-            [-73.63824144672192, 45.45886539033992],
-            [-73.63820370035012, 45.458849879455585],
-            [-73.63817822246807, 45.45888416924112],
-            [-73.63812722184346, 45.4588650615992],
-            [-73.63809891344295, 45.45889503527877],
-            [-73.63805925100608, 45.45888227130729],
-            [-73.63804178707059, 45.458911542591295],
-            [-73.63800221421994, 45.45890226150311],
-            [-73.63790797162682, 45.45903518437544],
-            [-73.63792945464428, 45.45905411295752],
-            [-73.63789159830482, 45.45908310537001],
-            [-73.6378753499016, 45.45907835433377],
-            [-73.63785235045175, 45.459105370774154],
-            [-73.63814312551322, 45.45921617475281],
-            [-73.63820764923148, 45.45912474376513],
-            [-73.63824947447044, 45.45914203753517],
-            [-73.63823410763928, 45.459161689254984],
-            [-73.63865919240125, 45.45930897424975],
-            [-73.63878575856276, 45.45916523682101],
-            [-73.63893471101957, 45.459219544549114],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Vanier Extension",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-73.63899474160336, 45.45882803291397],
-            [-73.63844069997971, 45.45861268764807],
-            [-73.638287219777, 45.45881201516468],
-            [-73.6388409933528, 45.45902411706979],
-            [-73.63899474160336, 45.45882803291397],
-          ],
-        ],
-      },
-    },
-  ],
-};
