--- ConcordiaMaps/data/markersData.js
@@ -1,504 +1,0 @@
-export const Building = [
-  // SGW Campus buildings latitude and longitude
-
-  {
-    name: "Webster Library",
-    coordinate: {
-      latitude: 45.4968158,
-      longitude: -73.5779337,
-    },
-  },
-  {
-    name: "B Annex",
-    coordinate: {
-      latitude: 45.49791,
-      longitude: -73.579441,
-    },
-  },
-  {
-    name: "CL Annex",
-    coordinate: {
-      latitude: 45.494228,
-      longitude: -73.579299,
-    },
-  },
-  {
-    name: "CI Annex",
-    coordinate: {
-      latitude: 45.497438,
-      longitude: -73.579946,
-    },
-  },
-  {
-    name: "D Annex",
-    coordinate: {
-      latitude: 45.497834,
-      longitude: -73.579282,
-    },
-  },
-  {
-    name: "EN Annex",
-    coordinate: {
-      latitude: 45.496864,
-      longitude: -73.579561,
-    },
-  },
-  {
-    name: "ER Building",
-    coordinate: {
-      latitude: 45.496362,
-      longitude: -73.580212,
-    },
-  },
-  {
-    name: "ES Building",
-    coordinate: {
-      latitude: 45.496172,
-      longitude: -73.579922,
-    },
-  },
-  {
-    name: "ET Building",
-    coordinate: {
-      latitude: 45.496163,
-      longitude: -73.579904,
-    },
-  },
-  {
-    name: "EV Building",
-    coordinate: {
-      latitude: 45.495538,
-      longitude: -73.577779,
-    },
-  },
-  {
-    name: "FA Annex",
-    coordinate: {
-      latitude: 45.496826,
-      longitude: -73.579481,
-    },
-  },
-  {
-    name: "Faubourg Tower",
-    coordinate: {
-      latitude: 45.494683,
-      longitude: -73.577659,
-    },
-  },
-  {
-    name: "Faubourg Building",
-    coordinate: {
-      latitude: 45.494362,
-      longitude: -73.578346,
-    },
-  },
-  {
-    name: "Grey Nuns Annex",
-    coordinate: {
-      latitude: 45.494071,
-      longitude: -73.577986,
-    },
-  },
-  {
-    name: "Guy-De Maisonneuve Building",
-    coordinate: {
-      latitude: 45.495682,
-      longitude: -73.578795,
-    },
-  },
-  {
-    name: "Grey Nuns Building - Main Entrance",
-    coordinate: {
-      latitude: 45.4935,
-      longitude: -73.576873,
-    },
-  },
-  {
-    name: "GS Building",
-    coordinate: {
-      latitude: 45.496331,
-      longitude: -73.580436,
-    },
-  },
-  {
-    name: "Henry F. Hall",
-    coordinate: {
-      latitude: 45.4973129,
-      longitude: -73.578876,
-    },
-  },
-  {
-    name: "K Annex",
-    coordinate: {
-      latitude: 45.497803,
-      longitude: -73.579357,
-    },
-  },
-  {
-    name: "LC Building",
-    coordinate: {
-      latitude: 45.496782,
-      longitude: -73.577358,
-    },
-  },
-  {
-    name: "LD Building",
-    coordinate: {
-      latitude: 45.496794,
-      longitude: -73.577231,
-    },
-  },
-  {
-    name: "Learning Square",
-    coordinate: {
-      latitude: 45.4963149,
-      longitude: -73.579472,
-    },
-  },
-  {
-    name: "M Annex",
-    coordinate: {
-      latitude: 45.497402,
-      longitude: -73.579715,
-    },
-  },
-  {
-    name: "John Molson School Of Business",
-    coordinate: {
-      latitude: 45.495309,
-      longitude: -73.579023,
-    },
-  },
-  {
-    name: "MI Annex",
-    coordinate: {
-      latitude: 45.490889,
-      longitude: -73.582412,
-    },
-  },
-  {
-    name: "MK Annex",
-    coordinate: {
-      latitude: 45.496606,
-      longitude: -73.579025,
-    },
-  },
-  {
-    name: "MM Annex",
-    coordinate: {
-      latitude: 45.494665,
-      longitude: -73.576315,
-    },
-  },
-  {
-    name: "MN Annex",
-    coordinate: {
-      latitude: 45.494568,
-      longitude: -73.579553,
-    },
-  },
-  {
-    name: "Montefiore Building, MT",
-    coordinate: {
-      latitude: 45.494442,
-      longitude: -73.576108,
-    },
-  },
-  {
-    name: "MU Annex",
-    coordinate: {
-      latitude: 45.497896,
-      longitude: -73.579553,
-    },
-  },
-  {
-    name: "P Annex",
-    coordinate: {
-      latitude: 45.496661,
-      longitude: -73.579165,
-    },
-  },
-  {
-    name: "PR Annex",
-    coordinate: {
-      latitude: 45.49697,
-      longitude: -73.579851,
-    },
-  },
-  {
-    name: "Q Annex",
-    coordinate: {
-      latitude: 45.496619,
-      longitude: -73.579122,
-    },
-  },
-  {
-    name: "R Annex",
-    coordinate: {
-      latitude: 45.496771,
-      longitude: -73.579424,
-    },
-  },
-  {
-    name: "RR Annex",
-    coordinate: {
-      latitude: 45.497407,
-      longitude: -73.579848,
-    },
-  },
-  {
-    name: "S Annex",
-    coordinate: {
-      latitude: 45.497423,
-      longitude: -73.579851,
-    },
-  },
-  {
-    name: "T Annex",
-    coordinate: {
-      latitude: 45.49671,
-      longitude: -73.57927,
-    },
-  },
-  {
-    name: "Samuel Bronfman Building, SB Building",
-    coordinate: {
-      latitude: 45.4966,
-      longitude: -73.58609,
-    },
-  },
-  {
-    name: "Toronto-Dominion Building",
-    coordinate: {
-      latitude: 45.494993,
-      longitude: -73.578129,
-    },
-  },
-  {
-    name: "TU Tunnel",
-    coordinate: {
-      latitude: 45.49648,
-      longitude: -73.578918,
-    },
-  },
-  {
-    name: "V Annex",
-    coordinate: {
-      latitude: 45.497021,
-      longitude: -73.579937,
-    },
-  },
-  {
-    name: "Visual Art Building",
-    coordinate: {
-      latitude: 45.495525,
-      longitude: -73.573794,
-    },
-  },
-  {
-    name: "X Annex",
-    coordinate: {
-      latitude: 45.496898,
-      longitude: -73.579722,
-    },
-  },
-  {
-    name: "Z Annex",
-    coordinate: {
-      latitude: 45.496931,
-      longitude: -73.579743,
-    },
-  },
-
-  // Loyola Campus buildings latitude and longitude
-
-  {
-    name: "Administration Building",
-    coordinate: {
-      latitude: 45.457772,
-      longitude: -73.639901,
-    },
-  },
-  {
-    name: "BB Annex",
-    coordinate: {
-      latitude: 45.459754,
-      longitude: -73.639176,
-    },
-  },
-  {
-    name: "BH Annex",
-    coordinate: {
-      latitude: 45.459729,
-      longitude: -73.639102,
-    },
-  },
-  {
-    name: "Central Building",
-    coordinate: {
-      latitude: 45.458271,
-      longitude: -73.64045,
-    },
-  },
-  {
-    name: "Communication Studies and Journalism Building",
-    coordinate: {
-      latitude: 45.457254,
-      longitude: -73.640404,
-    },
-  },
-  {
-    name: "Stinger Dome",
-    coordinate: {
-      latitude: 45.457656,
-      longitude: -73.636098,
-    },
-  },
-  {
-    name: "F.C. Smith Building",
-    coordinate: {
-      latitude: 45.458389,
-      longitude: -73.639291,
-    },
-  },
-  {
-    name: "Centre for Structural and Functional Genomics",
-    coordinate: {
-      latitude: 45.456849,
-      longitude: -73.640396,
-    },
-  },
-  {
-    name: "Hingston Hall, wing HA",
-    coordinate: {
-      latitude: 45.459471,
-      longitude: -73.641211,
-    },
-  },
-  {
-    name: "Hingston Hall, wing HB",
-    coordinate: {
-      latitude: 45.459162,
-      longitude: -73.641962,
-    },
-  },
-  {
-    name: "Hingston Hall, wing HC",
-    coordinate: {
-      latitude: 45.458504,
-      longitude: -73.643225,
-    },
-  },
-  {
-    name: "Applied Science Hub, HU Building",
-    coordinate: {
-      latitude: 45.458513,
-      longitude: -73.641921,
-    },
-  },
-  {
-    name: "Jesuit Residence",
-    coordinate: {
-      latitude: 45.458504,
-      longitude: -73.643225,
-    },
-  },
-  {
-    name: "PB Building",
-    coordinate: {
-      latitude: 45.456534,
-      longitude: -73.638106,
-    },
-  },
-  {
-    name: "Oscar Peterson Concert Hall, PT Building",
-    coordinate: {
-      latitude: 45.459308,
-      longitude: -73.638941,
-    },
-  },
-  {
-    name: "Perform Center",
-    coordinate: {
-      latitude: 45.45697,
-      longitude: -73.637255,
-    },
-  },
-  {
-    name: "Physical Service",
-    coordinate: {
-      latitude: 45.459694,
-      longitude: -73.639858,
-    },
-  },
-  {
-    name: "Psychology Building",
-    coordinate: {
-      latitude: 45.45905,
-      longitude: -73.640454,
-    },
-  },
-  {
-    name: "Recreational and Athletics",
-    coordinate: {
-      latitude: 45.456677,
-      longitude: -73.637523,
-    },
-  },
-  {
-    name: "Loyola Jesuit Hall and Conference Center",
-    coordinate: {
-      latitude: 45.458551,
-      longitude: -73.641016,
-    },
-  },
-  {
-    name: "Student Center",
-    coordinate: {
-      latitude: 45.459137,
-      longitude: -73.639203,
-    },
-  },
-  {
-    name: "Solar House, SH Building",
-    coordinate: {
-      latitude: 45.459298,
-      longitude: -73.642478,
-    },
-  },
-  {
-    name: "Terrebonne Building, TA Building",
-    coordinate: {
-      latitude: 45.460051,
-      longitude: -73.640842,
-    },
-  },
-  {
-    name: "TB Annex",
-    coordinate: {
-      latitude: 45.459992,
-      longitude: -73.640897,
-    },
-  },
-  {
-    name: "Richard J. Renaud Science Complex",
-    coordinate: {
-      latitude: 45.457567,
-      longitude: -73.641739,
-    },
-  },
-  {
-    name: "Vanier Library",
-    coordinate: {
-      latitude: 45.459058,
-      longitude: -73.638458,
-    },
-  },
-  {
-    name: "Vanier Extension",
-    coordinate: {
-      latitude: 45.459026,
-      longitude: -73.638606,
-    },
-  },
-];
